Bashed head against servo library and fixed some things, still non-functional

+++ robot-codebase/src/hardware/CustomServo.h
@@ -29,7 +29,7 @@ class CustomServo : public servo_interface
     unsigned long minPulse;
     unsigned long maxPulse;
     int pulseLength(int pos);
-    int dutyCycle(int length);
+    double dutyCycle(int length);
 
 };
 #endif //SERVO_H
